requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
python-dotenv==0.19.0
click==8.0.1
greenlet==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
SQLAlchemy==1.4.23
Werkzeug==2.0.1
psycopg2==2.9.1
requests==2.26.0
oauthlib==3.1.1
pyOpenSSL==20.0.1
Flask-Login==0.5.0
setuptools==58.0.4
waitress
flask-talisman
pytest